+++ validate/date/src/index.ts
@@ -1,0 +1,1 @@
+export { validateDate } from './validate.date.plugin'
